Add Whats new section for release 0.9.0
A few paragraphs of some of the key highlights
of the release and a pointer to more info.

Also calling out the Windows performance data
and addition of user documentation.

Updated with a date guess for August GA. Might
need tweaking after GA.

Finally, fixed the link to the Adopt binaries
in the first whats new entry because their
website has changed.

Closes: #81

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>