Updates for v0.10.0
Added a new section for 0.10.0 to the What's new page.

Updated the home page to remove references to OpenJDK
versions, tweak the download button URL target (releases
not nightlies) and do a little decluttering of links
(feedback from design review). Also added a couple of
words to direct people to the FAQ for what we support.

Removed references to IBM Knowledge Center now that our
user docs are growing.

Updated the FAQ support question to make it clearer
(OpenJDK versions and platforms)

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>