[WIP] Add a link to evidence that backs up perf claims
Feedback from research and others suggests that our
3 performance claims in blue are mistaken for
hyperlinks.

Users also wanted to see evidence, but missed finding
the performance page. Our link to performance is in
the "learn more" or can be found in the menu, but we
need to make it more obvious.

Changing the 3 performance claims to hyperlinks creates
a UX issue ... too many links joined together.

This PR to gain feedback on adding a 4th claim suggested
by Mark Stoodley, and adding a link to our evidence
"Show me the data!"

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>