Various fixes
Fixed indentation.

"Copy option" button was not working properly when the result contained reserved HTML characters - e.g. '<' and '>'. This was caused by mistakenly setting element.innerHTML instead of element.innerText on the temporary element used to store the text to be copied to the clipboard.

When using the tool agent, the range is set to 1..1 by default. Otherwise the default is 0..1. This is now reflected in the tool. I also decided to rework the range option slightly so that it is always enabled, but only appears in the result if it is modified from the default.

Implemented "opts" option for tool agent (ASYNC and WAIT).

Signed-off-by: Paul Cheeseman paul.cheeseman@uk.ibm.com