Various fixes
Fixed the indentation, which was a messy mixture of tabs and spaces.

"Copy option" button was not working properly when the result contained reserved HTML characters - e.g. '<' and '>'. This was caused by mistakenly setting element.innerHTML instead of element.innerText on the temporary element used to store the text to be copied to the clipboard.

Fixed a problem with the stack depth option not enabling/disabling correctly, which was caused by missing brackets on a call to isMethodJStackTraceActionSelected()

Signed-off-by: Paul Cheeseman paul.cheeseman@uk.ibm.com